Improvements to IP country data format
A few months ago, I introduced a new IP country data format using
Python files. While the intention was to find a balance between file
size, memory usage and performance, I failed to realize a few things:

- Generating the bytecode (.pyc) files uses 300+ MB memory momentarily.
  We don't want any sudden spikes in memory usage.
- The .pyc files are much larger than the .py files, ~8 MB vs ~3 MB.
  Some packages, such as the Arch package, ship multiple .pyc files
  for different optimization levels, resulting in ~16 MB of wasted
  space.
- Importing the data modules the first time a .pyc file is generated
  is quite slow.

This commit tries a different approach, using text files the same size
as the old .py files, without the need for bytecode files. This
fulfills the requirement of keeping the data files readable, which was
the primary motivation for changing the format in the first place.