Implement GTK 4 replacement for Gtk.EntryCompletion
Gtk.EntryCompletion is deprecated and doesn't support screen readers
in GTK 4.

Implement a replacement that combines the entry completion with the
combobox. Allows for reusing the model and menu popover.

Note that chat completion still uses Gtk.EntryCompletion for now.
We will most likely make the chat entry a subclass of ComboBox to
implement completion behavior specific to chats.